Add corrections to STARK section
There were not fundamental problem and it fact it was quite complete and concise. I elminated some part (specific times of the prover and verfier) because therey were not correct and we need to find the right values (if we want to talk about it).
I also eliminated the last paragraph because I don't agree. I might be wrong of